+++ lib/cpf_corrector.m
@@ -111,7 +111,7 @@ while (~converged && i < max_it)
     i = i + 1;
     
     %% evaluate Jacobian
-    [dSbus_dVm, dSbus_dVa] = dSbus_dV(Ybus, V);
+    [dSbus_dVa, dSbus_dVm] = dSbus_dV(Ybus, V);
     [dummy, neg_dSdb_dVm] = Sbusb(Vm);
     [dummy, neg_dSdt_dVm] = Sbust(Vm);
     dSbus_dVm = dSbus_dVm - neg_dSdb_dVm - lam * (neg_dSdt_dVm - neg_dSdb_dVm);